+++ libsofia-sip-ua/nta/sl_utils_log.c
@@ -267,8 +267,8 @@ void sl_payload_log(su_log_t *log, int level,
     log = su_log_default;
 
   while (s < end && *s != '\0') {
-    int n = strncspn(s, end - s, "\r\n");
-    int crlf = strnspn(s + n, end - s - n, "\r\n");
+    size_t n = strncspn(s, end - s, "\r\n");
+    size_t crlf = strnspn(s + n, end - s - n, "\r\n");
     if (n < 70) {
       memcpy(line, s, n);
       line[n] = '\0';
